requirements.txt
# Core Dependencies
fastapi>=0.104.0,<1.0.0
uvicorn[standard]>=0.24.0,<1.0.0
pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0

# Google Cloud Vertex AI (image generation)
google-cloud-aiplatform>=1.40.0,<2.0.0

# Supabase (storage and database)
supabase>=2.0.0,<3.0.0
postgrest>=0.13.0,<1.0.0

# Image Processing
Pillow>=10.0.0,<11.0.0
rembg>=2.0.50,<3.0.0  # Optional: Advanced background removal

# HTTP Client
httpx>=0.25.0,<1.0.0

# Utilities
python-multipart>=0.0.6  # For file uploads
python-jose[cryptography]>=3.3.0  # For JWT tokens
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0  # For TestClient

# Development
black>=23.0.0
ruff>=0.1.0
mypy>=1.7.0
